Token validation on /PutUserModel, /PostUserModel, /DeleteUserModel

+++ SET3-Backend/Controllers/UserModelsController.cs
@@ -63,14 +63,32 @@ namespace SET3_Backend.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<UserModel>> GetUserModel(int id)
         {
-            var userModel = await _context.UserModels.FindAsync(id);
-
-            if (userModel == null)
+            try
             {
-                return NotFound();
-            }
+                String token = Request.Cookies.Where(c => c.Key == "jwt").Select(c => c.Value).First();
 
-            return userModel;
+                if (token != null)
+                {
+                    if (ValidateToken(token) != null)
+                    {
+                        var userModel = await _context.UserModels.FindAsync(id);
+                        if (userModel == null)
+                        {
+                            return NotFound();
+                        }
+
+                        return userModel;
+                    }
+                    else
+                        return NoContent();
+                }
+            }
+            catch (Exception ex)
+            {
+                return NoContent();
+            }
+            return NoContent();
+
         }
 
         // PUT: api/UserModels/5
@@ -78,29 +96,44 @@ namespace SET3_Backend.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutUserModel(int id, UserModel userModel)
         {
-            if (id != userModel.Id)
-            {
-                return BadRequest();
-            }
-
-            _context.Entry(userModel).State = EntityState.Modified;
-
             try
             {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!UserModelExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
+                String token = Request.Cookies.Where(c => c.Key == "jwt").Select(c => c.Value).First();
 
+                if (token != null)
+                {
+                    if (ValidateToken(token) != null)
+                    {
+                        if (id != userModel.Id)
+                        {
+                            return BadRequest();
+                        }
+
+                        _context.Entry(userModel).State = EntityState.Modified;
+
+                        try
+                        {
+                            await _context.SaveChangesAsync();
+                        }
+                        catch (DbUpdateConcurrencyException)
+                        {
+                            if (!UserModelExists(id))
+                            {
+                                return NotFound();
+                            }
+                            else
+                            {
+                                throw;
+                            }
+                        }
+                    }
+                    else return NoContent();
+                }
+            }
+            catch (Exception ex)
+            {
+                return NoContent();
+            }
             return NoContent();
         }
 
@@ -109,26 +142,64 @@ namespace SET3_Backend.Controllers
         [HttpPost]
         public async Task<ActionResult<UserModel>> PostUserModel(UserModel userModel)
         {
-            _context.UserModels.Add(userModel);
-            await _context.SaveChangesAsync();
+            try
+            {
+                String token = Request.Cookies.Where(c => c.Key == "jwt").Select(c => c.Value).First();
 
-            return CreatedAtAction("GetUserModel", new { id = userModel.Id }, userModel);
+                if (token != null)
+                {
+                    if (ValidateToken(token) != null)
+                    {
+                        _context.UserModels.Add(userModel);
+                        await _context.SaveChangesAsync();
+                        return CreatedAtAction("GetUserModel", new { id = userModel.Id }, userModel);
+                    }
+                    else
+                        return NoContent();
+                }
+            }
+            catch (Exception ex)
+            {
+                return NoContent();
+            }
+            return NoContent();
         }
+
 
         // DELETE: api/UserModels/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUserModel(int id)
         {
-            var userModel = await _context.UserModels.FindAsync(id);
-            if (userModel == null)
+            try
             {
-                return NotFound();
+                String token = Request.Cookies.Where(c => c.Key == "jwt").Select(c => c.Value).First();
+
+                if (token != null)
+                {
+                    if (ValidateToken(token) != null)
+                    { 
+                        var userModel = await _context.UserModels.FindAsync(id);
+                        if (userModel == null)
+                        {
+                            return NotFound();
+                        }
+
+                        _context.UserModels.Remove(userModel);
+                        await _context.SaveChangesAsync();
+
+                        return NoContent();
+                    }
+                    else
+                        return NoContent();
+                }
             }
-
-            _context.UserModels.Remove(userModel);
-            await _context.SaveChangesAsync();
-
+            catch (Exception ex)
+            {
+                return NoContent();
+            }
             return NoContent();
+
+
         }
 
         private bool UserModelExists(int id)
